Core/Calendar: Fixed build, thx DoGoodS, close issue #34

+++ src/server/game/Entities/Creature/Creature.cpp
@@ -1721,7 +1721,7 @@ SpellInfo const *Creature::reachWithSpellAttack(Unit *victim)
         }
         if (bcontinue) continue;
 
-        if (spellInfo->ManaCost > GetPower(POWER_MANA))
+        if (spellInfo->ManaCost > (uint32)GetPower(POWER_MANA))
             continue;
         float range = spellInfo->GetMaxRange(false);
         float minrange = spellInfo->GetMinRange(false);
@@ -1764,7 +1764,7 @@ SpellInfo const *Creature::reachWithSpellCure(Unit *victim)
         }
         if (bcontinue) continue;
 
-        if (spellInfo->ManaCost > GetPower(POWER_MANA))
+        if (spellInfo->ManaCost > (uint32)GetPower(POWER_MANA))
             continue;
 
         float range = spellInfo->GetMaxRange(true);
